+++ java/src/grains_lxc/svm/My_SVM.java
@@ -4,7 +4,6 @@
  */
 package grains_lxc.svm;
 import grains_lxc.stat_model.StatModel;
-import grains_lxc.features.My_Features;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,7 +11,6 @@ import java.util.List;
 import org.opencv.core.Core;
 import org.opencv.core.CvType;
 import org.opencv.core.Mat;
-import org.opencv.core.Scalar;
 import org.opencv.ml.CvSVM;
 import org.opencv.ml.CvSVMParams;
 
@@ -35,14 +33,14 @@ public class My_SVM extends StatModel{
 		this.my_svm_params.set_gamma(svm_param_gamma);
 	}
 	
-	public void train(List<List<Float>> train_x, List<Float> train_y){
+	public void train(List<List<Double>> train_x, List<Double> train_y){
 		int item_features_qty=train_x.get(0).size();
 		int item_qty=train_y.size();
 		
 		Mat mat_train_x=new Mat(item_qty,item_features_qty,CvType.CV_32FC1);
 		Mat mat_train_y=new Mat(item_qty,1,CvType.CV_32FC1);
 		
-		// 填充数据到Mat类型中
+		// 填充数据到Mat类型中，将Double转成float型
 		for(int i=0;i<item_qty;i++){
 			mat_train_y.put(i,0,train_y.get(i).floatValue());
 			
@@ -54,14 +52,14 @@ public class My_SVM extends StatModel{
 		model_svm.train(mat_train_x, mat_train_y, new Mat(), new Mat(), my_svm_params);
 	}
 	
-	public List<Float>  predict(List<List<Float>> test_x){
+	public List<Double> predict(List<List<Double>> test_x){
 		int item_features_qty=test_x.get(0).size();
 		int item_qty=test_x.size();
 		
 		Mat mat_test_x=new Mat(item_qty,item_features_qty,CvType.CV_32FC1);
 		Mat mat_test_y=new Mat(item_qty,1,CvType.CV_32FC1);
 		
-		// 填充数据到Mat类型中
+		// 填充数据到Mat类型中，将Double转成float型
 		for(int i=0;i<item_qty;i++){
 			for(int j=0;j<item_features_qty;j++)
 				mat_test_x.put(i,j,test_x.get(i).get(j).floatValue());
@@ -69,9 +67,9 @@ public class My_SVM extends StatModel{
 		
 		model_svm.predict_all(mat_test_x, mat_test_y);
 				
-		List<Float> result=new ArrayList<Float>();
+		List<Double> result=new ArrayList<Double>();
 		for(int i=0;i<mat_test_y.rows();i++){
-			result.add((float)mat_test_y.get(i, 0)[0]);
+			result.add(mat_test_y.get(i, 0)[0]);
 		}
 		
 		return result;
